Got ToF sensors working with IO via VL53L0X library. Still need to test with all 8 sensors connected and still need to figure out how to set the addresses of each device. May involve a little tweaking to the VL53L0X library.

+++ Reveles/GUI/reveles_dbus_interface.h
@@ -58,8 +58,8 @@ Q_SIGNALS: // SIGNALS
     void ArduinoFound(bool good);
     void NucleoFound(bool good, int idx);
     void PIRStatus(bool stat, bool front);
-    void TOFReadings(int idx, float reading);
-    void USReadings(int idx, float reading);
+    void TOFReadings(int idx, double reading);
+    void USReadings(int idx, double reading);
     void servoUpdate(bool front, uint8_t dir);
     void motorUpdate(uint8_t dir);
 };
